* Improved networking code - should fix a deadlock which could occur during a DoS attack. * Improved NPC code - with better compatibility for buy with backpacks and ignore capacity. * Fixed several items.xml loading bugs. * Fixed a mailbox bug, and added "New mail has arrived" message. * Extended buffer size for look description to prevent overflows. * Fixed !createguild crash which would occur if database connection failed or triggers weren't properly imported. * Fixed player saving in SQLite. * Look description willl no longer add an extra dot if the text already ends with a dot, a question mark or an exclamation point. * Look description will now say "They weigh" for stackable items instead of "It weighs". * New Lua function: setCreatureLookDir(cid, dir). * Additional parameters to Lua function: doPlayerAddExp(cid, exp, <optional/NEW> usemultiplier, <optional/NEW> sendtext) * Monster code fixes. * Bumped version to 0.2.6. * Updated project files with version 0.2.6 and libboost_thread linker. * Code cleanup.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@2785 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ map.h
@@ -197,7 +197,7 @@ class Map
 		  * Set a single tile.
 		  * \param a tile to set for the position
 		  */
-		void setTile(uint16_t _x, uint16_t _y, uint16_t _z, Tile* newTile);
+		void setTile(int32_t _x, int32_t _y, int32_t _z, Tile* newTile);
 		void setTile(const Position& pos, Tile* newTile)
 		{
 			setTile(pos.x, pos.y, pos.z, newTile);
